Fixed small bugs

+++ src/main/java/bg/beesoft/beehive/service/BeehiveService.java
@@ -62,7 +62,7 @@ public class BeehiveService {
         Page<BeehiveView> beehiveViews = beehiveRepository.findByApiaryId(apiaryId, pageable)
                 .map(b -> modelMapper.map(b, BeehiveView.class));
 
-        chechApiaryAccess(userDetails, apiaryService.findById(apiaryId).getBeekeeper(), "Нямате достъп до този пчелин.");
+        checkApiaryAccess(userDetails, apiaryService.findById(apiaryId).getBeekeeper(), "Нямате достъп до този пчелин.");
 
         return beehiveViews;
     }
@@ -70,7 +70,7 @@ public class BeehiveService {
     public BeehiveFullView viewById(Long id, UserDetails userDetails) {
         BeehiveEntity beehiveEntity = findById(id);
 
-        chechApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
+        checkApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
 
         return modelMapper.map(beehiveEntity, BeehiveFullView.class);
     }
@@ -80,7 +80,7 @@ public class BeehiveService {
 
         BeehiveEntity beehiveEntity = findById(id);
 
-        chechApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
+        checkApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
 
         beehiveRepository.deleteById(id);
     }
@@ -88,13 +88,13 @@ public class BeehiveService {
     public Long findApiaryIdByBeehiveId(Long id, UserDetails userDetails) {
         BeehiveEntity beehiveEntity = findById(id);
 
-        chechApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
+        checkApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
 
         return beehiveEntity.getApiary().getId();
 
     }
 
-    private void chechApiaryAccess(UserDetails userDetails, UserEntity beekeeper, String s) {
+    private void checkApiaryAccess(UserDetails userDetails, UserEntity beekeeper, String s) {
         if (!beekeeper.getEmail().equals(userDetails.getUsername())) {
             throw new UnauthorizedRequestException(s);
         }
@@ -103,7 +103,7 @@ public class BeehiveService {
     public BeehiveEditDTO getEditDTOById(Long id, UserDetails userDetails) {
         BeehiveEntity beehiveEntity = findById(id);
 
-        chechApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
+        checkApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
 
         return modelMapper.map(beehiveEntity, BeehiveEditDTO.class);
     }
@@ -111,7 +111,7 @@ public class BeehiveService {
     public void updateBeehive(BeehiveEditDTO beehiveEditDTO, UserDetails userDetails, Long id) {
         BeehiveEntity beehiveEntity = findById(id);
 
-        chechApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
+        checkApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този пчелин.");
 
         ApiaryEntity apiaryEntity = apiaryService.findById(beehiveEditDTO.getApiaryId());
 
@@ -139,7 +139,7 @@ public class BeehiveService {
     public BeehiveEntity findById(Long id, UserDetails userDetails) {
         BeehiveEntity beehiveEntity = findById(id);
 
-        chechApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този кошер.");
+        checkApiaryAccess(userDetails, beehiveEntity.getBeekeeper(), "Нямате достъп до този кошер.");
         return beehiveEntity;
     }
 
@@ -148,7 +148,7 @@ public class BeehiveService {
     }
 
     public Page<BeehiveView> searchBeehives(SearchBeehiveDTO searchBeehiveDTO, Long apiaryId, Pageable pageable, UserDetails userDetails) {
-        chechApiaryAccess(userDetails, apiaryService.findById(apiaryId).getBeekeeper(), "Нямате достъп до този пчелин.");
+        checkApiaryAccess(userDetails, apiaryService.findById(apiaryId).getBeekeeper(), "Нямате достъп до този пчелин.");
 
         return beehiveRepository.findAll(new BeehiveSpecification(searchBeehiveDTO, apiaryId), pageable)
                 .map(b -> modelMapper.map(b, BeehiveView.class));
